# api/swagger.yaml
...
swagger: "2.0"
info:
  description: |
    Portainer API is an HTTP API served by Portainer. It is used by the Portainer UI and everything you can do with the UI can be done using the HTTP API.
    Examples are available at https://gist.github.com/deviantony/77026d402366b4b43fa5918d41bc42f8
    You can find out more about Portainer at [http://portainer.io](http://portainer.io) and get some support on [Slack](http://portainer.io/slack/).

    # Authentication

    Most of the API endpoints require to be authenticated as well as some level of authorization to be used.
    Portainer API uses JSON Web Token to manage authentication and thus requires you to provide a token in the **Authorization** header of each request
    with the **Bearer** authentication mechanism.

    Example:
    ```
    Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJhZG1pbiIsInJvbGUiOjEsImV4cCI6MTQ5OTM3NjE1NH0.NJ6vE8FY1WG6jsRQzfMqeatJ4vh2TWAeeYfDhP71YEE
    ```

    # Security

    Each API endpoint has an associated access policy, it is documented in the description of each endpoint.

    Different access policies are available:
    * Public access
    * Authenticated access
    * Restricted access
    * Administrator access

    ### Public access

    No authentication is required to access the endpoints with this access policy.

    ### Authenticated access

    Authentication is required to access the endpoints with this access policy.

    ### Restricted access

...

...

...

...

...

...

...

...

...
host: "portainer.domain"
basePath: "/api"
...
schemes:
...
paths:
...
  /endpoints:
    get:
      tags:
      - "endpoints"
      summary: "List endpoints"
      description: |
        List all endpoints based on the current user authorizations. Will
        return all endpoints if using an administrator account otherwise it will
        only return authorized endpoints.
        **Access policy**: restricted
      operationId: "EndpointList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/EndpointListResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "endpoints"
      summary: "Create a new endpoint"
      description: |
        Create a new endpoint that will be used to manage a Docker environment.
        **Access policy**: administrator
      operationId: "EndpointCreate"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "Name"
        in: "formData"
        type: "string"
        description: "Name that will be used to identify this endpoint (example: my-endpoint)"
        required: true
      - name: "EndpointType"
        in: "formData"
        type: "integer"
        description: "Environment type. Value must be one of: 1 (Docker environment), 2 (Agent environment) or 4 (Edge agent environment)"
        required: true
      - name: "URL"
        in: "formData"
        type: "string"
        description: "URL or IP address of a Docker host (example: docker.mydomain.tld:2375).\
          \ Defaults to local if not specified (Linux: /var/run/docker.sock, Windows: //./pipe/docker_engine)"
      - name: "PublicURL"
        in: "formData"
        type: "string"
        description: "URL or IP address where exposed containers will be reachable.\
          \ Defaults to URL if not specified (example: docker.mydomain.tld:2375)"
      - name: "GroupID"
        in: "formData"
        type: "string"
        description: "Endpoint group identifier. If not specified will default to 1 (unassigned)."
      - name: "TLS"
        in: "formData"
        type: "string"
        description: "Require TLS to connect against this endpoint (example: true)"
      - name: "TLSSkipVerify"
        in: "formData"
        type: "string"
        description: "Skip server verification when using TLS (example: false)"
      - name: "TLSSkipClientVerify"
        in: "formData"
        type: "string"
        description: "Skip client verification when using TLS (example: false)"
      - name: "TLSCACertFile"
        in: "formData"
        type: "file"
        description: "TLS CA certificate file"
      - name: "TLSCertFile"
        in: "formData"
        type: "file"
        description: "TLS client certificate file"
      - name: "TLSKeyFile"
        in: "formData"
        type: "file"
        description: "TLS client key file"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Endpoint"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Endpoint management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Endpoint management is disabled"
...
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  TLSConfiguration:
    type: "object"
    properties:
      TLS:
        type: "boolean"
        example: true
        description: "Use TLS"
      TLSSkipVerify:
        type: "boolean"
        example: false
        description: "Skip the verification of the server TLS certificate"
      TLSCACertPath:
        type: "string"
        example: "/data/tls/ca.pem"
        description: "Path to the TLS CA certificate file"
      TLSCertPath:
        type: "string"
        example: "/data/tls/cert.pem"
        description: "Path to the TLS client certificate file"
      TLSKeyPath:
        type: "string"
        example: "/data/tls/key.pem"
        description: "Path to the TLS client key file"
...
  Endpoint:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Endpoint identifier"
      Name:
        type: "string"
        example: "my-endpoint"
        description: "Endpoint name"
      Type:
        type: "integer"
        example: 1
        description: "Endpoint environment type. 1 for a Docker environment or 2 for an agent on Docker environment"
      URL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address of the Docker host associated to this endpoint"
      PublicURL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address where exposed containers will be reachable"
      GroupID:
        type: "integer"
        example: 1
        description: "Endpoint group identifier"
      AuthorizedUsers:
        type: "array"
        description: "List of user identifiers authorized to connect to this endpoint"
        items:
          type: "integer"
          example: 1
          description: "User identifier"
      AuthorizedTeams:
        type: "array"
        description: "List of team identifiers authorized to connect to this endpoint"
        items:
          type: "integer"
          example: 1
          description: "Team identifier"
      TLSConfig:
        $ref: "#/definitions/TLSConfiguration"
  EndpointSubset:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Endpoint identifier"
      Name:
        type: "string"
        example: "my-endpoint"
        description: "Endpoint name"
      Type:
        type: "integer"
        example: 1
        description: "Endpoint environment type. 1 for a Docker environment or 2 for an agent on Docker environment"
      URL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address of the Docker host associated to this endpoint"
      PublicURL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address where exposed containers will be reachable"
      GroupID:
        type: "integer"
        example: 1
        description: "Endpoint group identifier"
      AuthorizedUsers:
        type: "array"
        description: "List of user identifiers authorized to connect to this endpoint"
        items:
          type: "integer"
          example: 1
          description: "User identifier"
      AuthorizedTeams:
        type: "array"
        description: "List of team identifiers authorized to connect to this endpoint"
        items:
          type: "integer"
          example: 1
          description: "Team identifier"
      TLSConfig:
        $ref: "#/definitions/TLSConfiguration"
  GenericError:
    type: "object"
    properties:
      err:
        type: "string"
        example: "Something bad happened"
        description: "Error message"
...
  EndpointListResponse:
    type: "array"
    items:
      $ref: "#/definitions/EndpointSubset"
...